backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.3
SQLAlchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
python-multipart>=0.0.9
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
email-validator>=2.1

